Refactor dust data parsing
